refactor: extract DiagnosticsEventTypes.swift from DiagnosticsTypes.swift
Splits DiagnosticsTypes.swift into two files by concern:
- DiagnosticsTypes.swift keeps log-severity types
  (DiagnosticsLogCategory, DiagnosticsLogLevel)
- DiagnosticsEventTypes.swift gets event-naming types
  (DiagnosticsEventName + DiagnosticsEvent + extension)

Byte-identical move of lines 38-92. Both files import Foundation.
The extracted file no longer needs OSLog (only log-severity types did).

DiagnosticsTypes.swift: 92 → 36 lines. New file 57 lines. Tests: 667.
